Enhanced the existing text detection algorithm by testing and experimenting with different image examples and ER filtering parameters.
Implemented several filtering algorithms to reduce the amount of false detected sections
and to merge several text sections which are intersecting.

Manually tested the examples with this new approach and with tesseract
only with different page segmentation modes like (4, 6). It shows that
this approach is very good for some cases where tesseract shows
incorrect output, but also there are several examples where tesseract's
output is slightly better than this one.

Enhanced the testing mechanism and updated the output.json format which
now contains array of recognized text.